[ios] Always Select n-th Visible Index in Find in Page
Before this change, selectAndScrollToMatch() was selecting the n-th total
match, including invisible matches, and if the n-th match was not visible,
the next visible match was selected. In addition, it was returning the
total index value of the  selected match, which included invisible matches.
This is particularly problematic when the first overall match was not visible.
The logic would select the next visible match and return that match's overall index.
It could not return 0 because the logic has no idea how many visible matches precede
that index. To remedy these issues, always select the n-th
visible index and return that n-th visible index value.

Bug: 965834
Change-Id: I45607c18b8532132de304d230b6a9537c37c9f6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626658
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#664944}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1fd2c6f255ce9db4860b84a4d3a5a7d068087118