[omnibox] Query in Omnibox - Mark RELOAD page transition correctly
Previously, when the search query was displayed in the omnibox
(i.e. Query in Omnibox was active), the user focused the omnibox, and
pressed Enter without changing the text, the page transition would
be marked as GENERATED, i.e. from a search suggestion.

In reality, these are more akin to RELOAD.

We already have special logic to handle URL navigations without text
edits. This just extends that logic to Query in Omnibox too.

Although the logic is duplicated for Android vs. all other platforms,
it seemed like too big of a job to de-duplicate this logic in this CL.

Bug: 963554
Change-Id: I1e0d8798123b12e2ed0bc50df2541ed056a90232
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1641633
Commit-Queue: Tommy Li <tommycli@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#666909}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fddf70e1476e07e4e8b5ac03285406a2f79dd197