[iOS] Fix traitCollection in iOS 13
In iOS 13, -traitCollectionDidChange is no longer called when the view
is added to the view hierarchy. In previous version this was called
and some code was written with that assumption in mind.
This CL fixes the calls where needed.

Bug: 973033
Change-Id: I563e20b00eac14714895f4d6c3674764f3a3039c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1655448
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#668383}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c4b8f13dca2336213880ae5996afd61cc7cca303