[ios] Only add Infobar Badge as an accessibility item if shown
Fixes a bug where an accessibility element would be present for the
Infobar Badge even when it was not visible. This change adds and
removes the badge as an accessibility element when the badge is
being displayed/hidden.

Change-Id: I618ee3f7b684cb479fc70fbafbd329b99e27e18c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1652045
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#668962}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6841ceba55d970693035f4a7a23dd8fbf9b7d5ad