Remove unnecessary includes from IdentityManager's public header
The include for accounts_cookie_mutator.h is not really needed since we
already have a forward declaration for AccountsCookieMutator, so let's
remove it and add it explicitly to inline_login_handler_impl.cc.

As for account_fetcher_service.h, we can use a forward declaration for
this one as well and move the include to the implementation file, so
that we eliminate this transitive include as well, so let's do it.

Bug: 939372
Change-Id: I55164145021f32fc3e1d88fcc9372c3881482326
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1660444
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Lowell Manners <lowell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#669628}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0294543adfaaea3babfc81066992dfee2f88a9f2