[iOS] Fix iOS 13 layout guide issues
On iOS 13, the traitCollection is correct when the buttons are created.
The value of hiddenInCurrentSizeClass is correct when the button is
created, before it is added to the view hierarchy. In that case, the
button isn't able to find its associated named guide.
When the button is added to the view hierarchy later, the value of
hiddenInCurrentSizeClass doesn't change so the named guide isn't
checked, preventing it to be correctly associated with the button.

This CL fixes it by always checking if the layout guide is correctly
associated.

Bug: 971730
Change-Id: I3f058e1a97c902bad0e080ae17f5a516ddcb6985
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1669294
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#670893}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 576e072dfe226f6ef7cd3ca0a0030a19b5f1d9a5