Suppress testInfobarAlwaysTranslate failure on iOS 13.
On iOS 13 "Always Translate French" item is not visible due to
a real bug in the options menu.

This CL relaxes matching requirement for iOS 13 to make the test pass.
The real fix will be landed separately and this workaround will be
removed.

Bug: 979079
Change-Id: Icfbd6a39f1ca7ff93fdc7dbad59414004232503b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1679637
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#673385}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 90cf93ea04b82e962948ad6d3c2181db1adc8389