[ios]Stops using selected cell property for adding translate checkmark
UITableViewCell selected property should be used to set a cell state
to selected. I believe this shouldn't be used to set some styling that
is not related to the actual "selection" of UITableViewCells in a UITableView
context e.g. Selecting cells will call tableView:didSelectRowAtIndexPath and
then these can be deleted.

On iOS13 this was causing a strange bug where some cells were on top of
each other.

Since selected was only being used to display a checkmark or not, moving
this to a stand alone method fixes the issue.

Bug: 979079
Change-Id: I797b7dd3867896cc640a78854f1d06cb29b38f03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1684248
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#674204}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dabf1ca153d6e92b071cb937fe333c59ba9d001d