[iOS] Fix Omnibox gray text in steady state
The gray text came from the placeholder font color. According to the
documentation, setting the |attributedText| property of a UILabel
overrides the |textColor| with the color of the first character of
the attributed string. However, this bug appears only on iOS-13, so
that may be a recent change.

Bug: 979202
Change-Id: Ic387864629d92a4e0cefb42dcf60dd2a58c076c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1689199
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#675580}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec70e2e7d7e38ab1a7f9446680b5e2b4c49d0611