Fix saving password hash on Chrome sign-in.
The problem was that after a successful Chrome sign-in, a tab navigates
to the new tab page which is processed in different way than web
pages. That CL introduces saving a password hash in case when the tab
navigates to the new tab page after Chrome sign-in.

Bug: 979046

Change-Id: Idd030e32876647b4ed995ff8fbda617697820774
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1690844
Commit-Queue: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#675774}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 43d8751518a7bdfacc85c4b89647c7f0b3c301d8