Remove hosted_domain parameter to AuthenticationService::SignIn
Remove the parameter as it is not used (except to check that the
value is correctly set for the primary account). Instead check
that a non-empty value is set for the primary account after the
call to DeviceAccountsSynchronizer::ReloadAllAccountsFromSystem.

Bug: 982681
Change-Id: I9627ef85e5e1f4df66872418ea8ac07cc72f03b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695307
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#676430}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1bfb9c36d356d70c4758339b3109205e116a60b6