Disable StopNavigationAfterPolicyDeciderCallback for iOS 13.
WKWebView has undefined behavior regarding start the navigation
on iOS 13. It's not a bug, just a difference in behavior from
iOS 11 and 12, where navigation always starts on simulator but
never starts on device.

The test is disabled for iOS 13, because it's not possible
(and unnecessary) to test undefined behavior.

Bug: 983150
Change-Id: I69f5b503bc994bebd803fda4021cb88d4ba0c3e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1698779
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#676878}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 266ef855eaa06898bbe69459937a9db1a205eb06