Allow ZPS with empty URL if mechanism requires no URL.
This change fixes a bug, where page classification fails to detect NTP.
Simplest scenario leading to this case is navigating to a URL and then
using home button, bringing user back to a NTP, for which URL is not
meant to be displayed.

Page classification mechanism failed in this case to correctly identify
the page as a NTP and fell back to INVALID, since the URLs returned by
calls to GetURL() were empty (instead of chrome://newtab).

Bug: 978660
Change-Id: I45d350571d90ee8d4babbfb0eee19377f7d56711
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1676999
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Ender <ender@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#677810}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d739c811d0eac6380b3b09d0e4155eca35e08f48