[iOS] Use dark mode colors for BVC toolbars in incogito
On iOS 13, we can achieve this by using the overrideUserInterfaceStyle
property to force the toolbars into dark styling. Then, all the UI
elements can be the same UIColor in both modes (because that color is
dynamic).

This does not fix the problem for iOS 12, which will come in a later
CL.

Bug: 976643
Change-Id: I76eb50f6484f7ac9a846fb8cb50e5ffc46c0f8a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709322
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#679546}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2d7b878de459aaeb886c46f2812414120790bcf1